+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - (Bugfix) Do not tolerate False Bootstrap condition in UpToDate evaluation
 - (Improvement) Don't serialize and deprecate two DeploymentReplicationStatus fields
 - (Improvement) Improve error message when replication can't be configured
+- (Bugfix) Fix License handling in case of broken license secret
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress